Correct transfer data queue to server

+++ 10-projects/rover_rasp/rover_system/lib/include/shared_memory.h
@@ -9,7 +9,7 @@ typedef struct shared_memory_st{
 }shared_memory_st;
 
 typedef struct generic{
-  long int id;
+  int id;
   char command[512];
 }generic_st;
